requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
black==22.1.0
blue==0.9.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
django-environ==0.12.0
exceptiongroup==1.3.0
flake8==4.0.1
gunicorn==23.0.0
iniconfig==2.1.0
kombu==5.5.4
mccabe==0.6.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
redis==6.3.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
